pyproject.toml
[tool.poetry]
name = "smart-kit"
version = "0.1.0"
description = ""
authors = ["igor"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
alembic = "^1.16.1"
sqlalchemy = "^2.0.41"
pydantic-settings = "^2.9.1"
asyncpg = "^0.30.0"
pyjwt = "^2.10.1"
httpx = "^0.28.1"
bcrypt = "^4.3.0"
aioredis = "^2.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.11"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
